requirements.txt
async-timeout==4.0.2
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.5.5
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
